pyproject.toml
# pyproject.toml

## ComfyUI Config and Info
[project]
name =  "comfyui-structured-outputs"
version = "0.1.0"
requires-python = ">=3.10"
description = "ComfyUI nodes for (V)LLM Structured Outputs"
readme = "README.md"
license = {text = "MIT License"}
# use requirements.in to manage dependencies, don't directly update here, see README.md
dependencies = [
    "aiohappyeyeballs==2.4.4",
    "aiohttp==3.11.11",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "anyio==4.8.0",
    "attrs==25.1.0",
    "certifi==2024.12.14",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "colorama==0.4.6 ; sys_platform == 'win32'",
    "distro==1.9.0",
    "docstring-parser==0.16",
    "filelock==3.17.0",
    "frozenlist==1.5.0",
    "fsspec==2024.12.0",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httpx==0.28.1",
    "idna==3.10",
    "instructor==1.7.2",
    "jinja2==3.1.5",
    "jiter==0.8.2",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "mpmath==1.3.0",
    "multidict==6.1.0",
    "networkx==3.4.2",
    "numpy==2.2.2",
    "nvidia-cublas-cu12==12.4.5.8 ; platform_machine == 'x86_64' and sys_platform == 'linux'",
    "nvidia-cuda-cupti-cu12==12.4.127 ; platform_machine == 'x86_64' and sys_platform == 'linux'",
    "nvidia-cuda-nvrtc-cu12==12.4.127 ; platform_machine == 'x86_64' and sys_platform == 'linux'",
    "nvidia-cuda-runtime-cu12==12.4.127 ; platform_machine == 'x86_64' and sys_platform == 'linux'",
    "nvidia-cudnn-cu12==9.1.0.70 ; platform_machine == 'x86_64' and sys_platform == 'linux'",
    "nvidia-cufft-cu12==11.2.1.3 ; platform_machine == 'x86_64' and sys_platform == 'linux'",
    "nvidia-curand-cu12==10.3.5.147 ; platform_machine == 'x86_64' and sys_platform == 'linux'",
    "nvidia-cusolver-cu12==11.6.1.9 ; platform_machine == 'x86_64' and sys_platform == 'linux'",
    "nvidia-cusparse-cu12==12.3.1.170 ; platform_machine == 'x86_64' and sys_platform == 'linux'",
    "nvidia-nccl-cu12==2.21.5 ; platform_machine == 'x86_64' and sys_platform == 'linux'",
    "nvidia-nvjitlink-cu12==12.4.127 ; platform_machine == 'x86_64' and sys_platform == 'linux'",
    "nvidia-nvtx-cu12==12.4.127 ; platform_machine == 'x86_64' and sys_platform == 'linux'",
    "openai==1.60.2",
    "pillow==11.1.0",
    "propcache==0.2.1",
    "pydantic==2.10.6",
    "pydantic-core==2.27.2",
    "pygments==2.19.1",
    "python-dotenv==1.0.1",
    "requests==2.32.3",
    "rich==13.9.4",
    "setuptools==75.8.0",
    "shellingham==1.5.4",
    "sniffio==1.3.1",
    "sympy==1.13.1",
    "tenacity==9.0.0",
    "torch==2.5.1",
    "torchvision==0.20.1",
    "tqdm==4.67.1",
    "triton==3.1.0 ; python_full_version < '3.13' and platform_machine == 'x86_64' and sys_platform == 'linux'",
    "typer==0.15.1",
    "typing-extensions==4.12.2",
    "urllib3==2.3.0",
    "yarl==1.18.3",
]

[project.urls]
Repository = "https://github.com/tigeryy2/comfyui-structured-outputs"

[tool.comfy]
PublisherId = "tigeryy2"
DisplayName = "ComfyUI Structured Outputs" # Display name for the Custom Node. Can be changed later.
Icon = "https://example.com/icon.png" # SVG, PNG, JPG or GIF (MAX. 800x400px)

## Other Config
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
# add scripts here that are installed with the environment
# uses 'uv pip compile requirements.in > requirements.txt' to compile all deps
compile-reqs = "python_template.utils.scripts.generate_requirements_txt:main"

[tool.setuptools.packages.find]
where = ["."]
exclude = ["tests", "docs"]

[tool.uv.pip]
universal = true    # UV to generate universal lockfiles

[tool.uv.sources]
python-template = { workspace = true }
comfyui-structured-outputs = { workspace = true }

[tool.ruff]
extend-exclude = []
show-fixes = true
target-version = "py312"

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
ignore = [
    "E501",  # line too long
    "B008",  # do not perform calls in argument defaults
]

[tool.pytest.ini_options]
addopts = "--html=logs/report.html --self-contained-html"
testpaths = ["tests"]
log_cli = "True"
log_cli_level = "INFO"
log_cli_format = "[%(asctime)s] [%(levelname)s] %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
log_file = "./logs/pytest.log"

[dependency-groups]
# use requirements-dev.in to manage dev dependencies, don't directly update here
dev = [
    "anyio==4.8.0",
    "black==24.10.0",
    "certifi==2024.12.14",
    "cfgv==3.4.0",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "colorama==0.4.6 ; sys_platform == 'win32'",
    "comfy-cli==1.3.5",
    "coverage==7.6.10",
    "distlib==0.3.9",
    "filelock==3.17.0",
    "gitdb==4.0.12",
    "gitpython==3.1.44",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httpx==0.28.1",
    "identify==2.6.6",
    "idna==3.10",
    "iniconfig==2.0.0",
    "jinja2==3.1.5",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "mixpanel==4.10.1",
    "mypy-extensions==1.0.0",
    "nodeenv==1.9.1",
    "packaging==24.2",
    "pathspec==0.12.1",
    "platformdirs==4.3.6",
    "pluggy==1.5.0",
    "pre-commit==4.1.0",
    "prompt-toolkit==3.0.50",
    "psutil==6.1.1",
    "pygments==2.19.1",
    "pytest==8.3.4",
    "pytest-cov==6.0.0",
    "pytest-html==4.1.1",
    "pytest-metadata==3.1.1",
    "pyyaml==6.0.2",
    "questionary==2.1.0",
    "requests==2.32.3",
    "rich==13.9.4",
    "ruff==0.9.3",
    "semver==3.0.4",
    "setuptools==75.8.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "tomlkit==0.13.2",
    "typer==0.15.1",
    "typing-extensions==4.12.2",
    "urllib3==2.3.0",
    "uv==0.5.25",
    "virtualenv==20.29.1",
    "wcwidth==0.2.13",
    "websocket-client==1.8.0",
    "wheel==0.45.1",
]
